create LinkStack by input data and replace all ID to Int type

+++ genogram/src/main/kotlin/ffc/genogram/FamilyTreeDrawer.kt
@@ -164,7 +164,7 @@ class FamilyTreeDrawer {
         return 0
     }
 
-    fun findPersonIndById(personId: Long, layerNumb: Int): Int {
+    fun findPersonIndById(personId: Int, layerNumb: Int): Int {
         val nodeList = personFamilyStorage[layerNumb]
 
         nodeList.forEachIndexed { index, person ->
@@ -176,7 +176,7 @@ class FamilyTreeDrawer {
         return 0
     }
 
-    fun getPersonIndById(personId: Long, layerNumb: Int): Person? {
+    fun getPersonIndById(personId: Int, layerNumb: Int): Person? {
         val nodeList = personFamilyStorage[layerNumb]
 
         nodeList.forEachIndexed { index, person ->
@@ -247,7 +247,7 @@ class FamilyTreeDrawer {
 
     fun addMarriageLineInd(layerNumb: Int, focusedPerson: Person, addedPerson: Person?): Int {
         // Check the person gender, and find the person's
-        val allPeople: MutableList<Long> = mutableListOf()
+        val allPeople: MutableList<Int> = mutableListOf()
         personFamilyStorage[layerNumb - 1].forEach {
             if (it is Person)
                 allPeople.add(it.idCard)
